[MOV] purchase_alternative*: separate models files
*: purchase_requisition

With This Commit:
=================================

We have separated the code that depends on the alternative logic
of models files from the purchase_requisition module and moved them
into the purchase alternative module.

taskId :- 4737059
