+++ addons/purchase_alternative/models/__init__.py
@@ -1,5 +1,3 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import purchase
-from . import product
-from . import purchase_requisition
